Move handling of the "confirmed" flag into application service layer

+++ src/Service/DistrictService.php
@@ -85,6 +85,9 @@ class DistrictService
 
     public function remove(RemoveDistrictCommand $command): void
     {
+        if (!$command->isConfirmed()) {
+            return;
+        }
         $this->districtRepository->remove($this->get($command->getId()));
     }
